added basic file conversion for pipes. Refactored filestructure.
