--- gbk_test.go
@@ -1,15 +1,0 @@
-package main
-
-import "testing"
-
-func BenchmarkReadGbk(b *testing.B) {
-	for i := 0; i < b.N; i++ {
-		ReadGbk("data/bsub.gbk")
-	}
-}
-
-func BenchmarkReadGbk1(b *testing.B)     { BenchmarkReadGbk(b) }
-func BenchmarkReadGbk10(b *testing.B)    { BenchmarkReadGbk(b) }
-func BenchmarkReadGbk100(b *testing.B)   { BenchmarkReadGbk(b) }
-func BenchmarkReadGbk1000(b *testing.B)  { BenchmarkReadGbk(b) }
-func BenchmarkReadGbk10000(b *testing.B) { BenchmarkReadGbk(b) }
